refactor: create mutations and queries folder for hooks related to api
