+++ src/hooks/api/mutations/useUpdateUserAvatar/index.ts
