+++ src/hooks/api/queries/useUserAvatarURI/index.ts
